refactor: colocate quest create modal tests
